requirements.txt
aioredis==1.3.1
asgiref==3.2.10
async-timeout==3.0.1
attrs==19.3.0
autobahn==20.6.2
Automat==20.2.0
beautifulsoup4==4.9.1
bs4==0.0.1
certifi==2020.6.20
cffi==1.14.0
channels==2.4.0
channels-redis==3.0.0
chardet==3.0.4
constantly==15.1.0
cryptography==2.9.2
daphne==2.5.0
dataclasses==0.7
Django==3.0.5
django-cors-headers==3.4.0
djangorestframework==3.11.0
djongo==1.3.3
dnspython==1.16.0
gunicorn==20.0.4
hiredis==1.0.1
httpie==2.2.0
hyperlink==19.0.0
idna==2.10
incremental==17.5.0
joblib==0.16.0
msgpack==1.0.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.20
Pygments==2.6.1
PyHamcrest==2.0.2
pyOpenSSL==19.1.0
pytz==2020.1
requests==2.24.0
service-identity==18.1.0
six==1.15.0
soupsieve==2.0.1
sqlparse==0.2.4
threadpoolctl==2.1.0
Twisted==20.3.0
txaio==20.4.1
urllib3==1.25.9
whitenoise==5.1.0
zope.interface==5.1.0
